+++ leviathan/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-__version__ = "2025.3.11"
+__version__ = "2025.3.12"
 from . import utils
 from . import index