+++ leviathan/profile_pathway.py
@@ -4,6 +4,7 @@ from collections import defaultdict
 from tqdm import tqdm
 import pandas as pd
 import numpy as np
+import xarray as xr
 # from memory_profiler import profile
 
 from kegg_pathway_profiler.pathways import (
@@ -209,3 +210,218 @@ def aggregate_feature_abundance_for_clusters(df_feature_abundance:pd.DataFrame, 
     df_output = df_feature_abundance.groupby(f).sum()
     df_output.index = pd.MultiIndex.from_tuples(df_output.index, names=["id_genome_cluster", "id_feature"])
     return df_output
+
+def merge_pathway_profiling_tables_as_pandas(profiling_directory:str, data_type:str, level="genomes", metric="number_of_reads", fillna_with_zeros:bool=False, sparse:bool=False):
+    
+    """
+    merges sample-level {data_type} values from multiple samples into a single DataFrame.
+
+    Parameters
+    ----------
+    profiling_directory : str
+        Path to directory containing sample-level directories with output files.
+    data_type : str
+        Type of {level}-level data to merge. One of: {"feature_abundances", "feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio", "gene_abundances", "pathway_abundances"}
+    level : str, optional
+        Level of organization for {data_type}. One of {"genomes", "genome_cluster"}.
+    metric : str, optional
+        Metric to use for {data_type}. One of {"number_of_reads", "tpm", "coverage"}.
+    fillna_with_zeros : bool, optional
+        Whether to fill missing values with zeros. Default is False.
+    sparse : bool, optional
+        Whether to return a pd.Sparse type. Default is False. 
+    Returns
+    -------
+    pd.DataFrame
+        merged DataFrame with {data_type} values for each sample.
+
+    Notes
+    -----
+    Will raise a ValueError if an invalid combination of arguments is provided, such as level="genome_cluster" and data_type="gene_abundances".
+    
+    Files:
+    * feature_abundances.genome_clusters.tsv.gz
+    * feature_abundances.genomes.tsv.gz
+    * feature_prevalence-binary.genome_clusters.tsv.gz
+    * feature_prevalence-binary.genomes.tsv.gz
+    * feature_prevalence.genome_clusters.tsv.gz
+    * feature_prevalence.genomes.tsv.gz
+    * feature_prevalence-ratio.genome_clusters.tsv.gz
+    * gene_abundances.genomes.tsv.gz
+    * pathway_abundances.genome_clusters.tsv.gz
+    * pathway_abundances.genomes.tsv.gz
+    """
+
+    check_argument_choice(
+        query=data_type, 
+        choices={"feature_abundances", "feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio", "gene_abundances", "pathway_abundances"},
+        )
+    check_argument_choice(
+        query=level, 
+        choices={"genomes", "genome_clusters"},
+        )
+    check_argument_choice(
+        query=metric, 
+        choices={"number_of_reads", "tpm", "coverage"},
+        )
+
+    illegal_conditions = [
+        (level == "genome_cluster") and (data_type == "gene_abundances"),
+        (level == "genomes") and (data_type == "feature_prevalence-ratio"),
+        (data_type != "pathway_abundances") and (metric == "coverage"),
+    ]
+    
+    if any(illegal_conditions):
+        raise ValueError(f"Invalid combination of arguments: level={level}, data_type={data_type}, metric={metric}")
+    
+    # Merge tables to produce output
+    filepaths = glob.glob(f"{profiling_directory}/*/output/{data_type}.{level}.tsv.gz")
+    if filepaths:
+        output = dict()
+        # Abundance/Coverage
+        if data_type in {"feature_abundances", "gene_abundances", "pathway_abundances"}:
+            
+            # Determine column name
+            column = str(metric)
+            if data_type in {"feature_abundances", "pathway_abundances"}:
+                if metric != "coverage":
+                    column = f"{metric}(scaled)"
+            
+            description = "Merging {}-level {} {} values".format(level, data_type.replace("_", " "), metric)
+            for filepath in tqdm(filepaths, description):
+                id_sample = filepath.split("/")[-3]
+                df = pd.read_csv(filepath, sep="\t", index_col=[0,1])
+                output[id_sample] = df[column]
+                
+        # Prevalence
+        elif data_type in {"feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio"}:
+            description = "Merging {}-level {} values".format(level, data_type.replace("_", " "))
+            for filepath in tqdm(filepaths, description):
+                id_sample = filepath.split("/")[-3]
+                df = pd.read_csv(filepath, sep="\t", index_col=0)
+                output[id_sample] = df.stack()
+        X = pd.DataFrame(output).T
+        
+        sparse_dtype = "float"
+        missing_value_fill = pd.NA
+        if data_type == "feature_prevalence-binary":
+            if fillna_with_zeros:
+                X = X.fillna(0)
+                sparse_dtype = "int"
+                missing_value_fill = 0
+        else:
+            if fillna_with_zeros:
+                X = X.fillna(0.0)
+                missing_value_fill = 0.0
+        if sparse:
+            X = X.astype(pd.SparseDtype(sparse_dtype, missing_value_fill))
+        return X
+                
+    else:
+        raise FileNotFoundError(f"Could not find any {data_type}.{level}.tsv.gz files in {profiling_directory}")
+
+def merge_pathway_profiling_tables_as_xarray(profiling_directory:str, data_type:str, level="genomes", metric="number_of_reads", fillna_with_zeros:bool=False):
+    
+    """
+    merges sample-level {data_type} values from multiple samples into a single DataFrame.
+
+    Parameters
+    ----------
+    profiling_directory : str
+        Path to directory containing sample-level directories with output files.
+    data_type : str
+        Type of {level}-level data to merge. One of: {"feature_abundances", "feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio", "gene_abundances", "pathway_abundances"}
+    level : str, optional
+        Level of organization for {data_type}. One of {"genomes", "genome_cluster"}.
+    metric : str, optional
+        Metric to use for {data_type}. One of {"number_of_reads", "tpm", "coverage"}.
+    fillna_with_zeros : bool, optional
+        Whether to fill missing values with zeros. Default is False.
+
+    Returns
+    -------
+    xr.DataArray
+        Merged xr.DataArray with dims: (samples, {level}, features)
+
+    Notes
+    -----
+    Will raise a ValueError if an invalid combination of arguments is provided, such as level="genome_cluster" and data_type="gene_abundances".
+    
+    Files:
+    * feature_abundances.genome_clusters.tsv.gz
+    * feature_abundances.genomes.tsv.gz
+    * feature_prevalence-binary.genome_clusters.tsv.gz
+    * feature_prevalence-binary.genomes.tsv.gz
+    * feature_prevalence.genome_clusters.tsv.gz
+    * feature_prevalence.genomes.tsv.gz
+    * feature_prevalence-ratio.genome_clusters.tsv.gz
+    * gene_abundances.genomes.tsv.gz
+    * pathway_abundances.genome_clusters.tsv.gz
+    * pathway_abundances.genomes.tsv.gz
+    """
+
+    check_argument_choice(
+        query=data_type, 
+        choices={"feature_abundances", "feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio", "gene_abundances", "pathway_abundances"},
+        )
+    check_argument_choice(
+        query=level, 
+        choices={"genomes", "genome_clusters"},
+        )
+    check_argument_choice(
+        query=metric, 
+        choices={"number_of_reads", "tpm", "coverage"},
+        )
+
+    illegal_conditions = [
+        (level == "genome_cluster") and (data_type == "gene_abundances"),
+        (level == "genomes") and (data_type == "feature_prevalence-ratio"),
+        (data_type != "pathway_abundances") and (metric == "coverage"),
+    ]
+    
+    if any(illegal_conditions):
+        raise ValueError(f"Invalid combination of arguments: level={level}, data_type={data_type}, metric={metric}")
+    
+    # Merge tables to produce output
+    filepaths = glob.glob(f"{profiling_directory}/*/output/{data_type}.{level}.tsv.gz")
+    if filepaths:
+        output = dict()
+        # Abundance/Coverage
+        if data_type in {"feature_abundances", "gene_abundances", "pathway_abundances"}:
+            
+            # Determine column name
+            column = str(metric)
+            if data_type in {"feature_abundances", "pathway_abundances"}:
+                if metric != "coverage":
+                    column = f"{metric}(scaled)"
+            
+            description = "Merging {}-level {} {} values".format(level, data_type.replace("_", " "), metric)
+            for filepath in tqdm(filepaths, description):
+                id_sample = filepath.split("/")[-3]
+                df = pd.read_csv(filepath, sep="\t", index_col=[0,1])
+                df = df[column].unstack()
+                output[id_sample] = xr.DataArray(data = df.values, coords = [(level, df.index), ("features", df.columns)])
+                del df
+                
+        # Prevalence
+        elif data_type in {"feature_prevalence", "feature_prevalence-binary", "feature_prevalence-ratio"}:
+            description = "Merging {}-level {} values".format(level, data_type.replace("_", " "))
+            for filepath in tqdm(filepaths, description):
+                id_sample = filepath.split("/")[-3]
+                df = pd.read_csv(filepath, sep="\t", index_col=0)
+                output[id_sample] = xr.DataArray(data = df.values, coords = [(level, df.index), ("features", df.columns)])
+        X = xr.concat(output.values(), dim="samples")
+        X["samples"] = list(output.keys())
+        
+        
+        if data_type == "feature_prevalence-binary":
+            if fillna_with_zeros:
+                X = X.fillna(0)
+            X = X.astype(int)
+        else:
+            if fillna_with_zeros:
+                X = X.fillna(0.0)
+        return X
+                
+    else:
+        raise FileNotFoundError(f"Could not find any {data_type}.{level}.tsv.gz files in {profiling_directory}")